Improve the setup process.
Update the setup.py process to automatically build the scala assembly.
The setup process checks for the existance of 'sbt', and runs the sbt
'clean' and 'assembly' commands during `build`.

Setup also copies the built jar file into the system classpath during
`install`, so that no classpath environment needs to be set.