Improved setup to be more reliable as a dependency.
Egg, wheel, and any bdist package won't install the companion JAR file.
The only method that seems to work is distributing a source-only
tarball, and running pip install on that tarball. Otherwise, some build
steps are skipped, or data files aren't installed, etc.

Additionally, the installation now follows the pattern from py4j for the
location of the jar file, installing it in the system "share" directory.